Dependency cycle fix
